# docs/clinic-v2/03_Architecture_and_Database/contracts/openapi.yaml
openapi: 3.1.1
info:
  title: Clinic Management V2 - Representative High-Risk Contracts
  version: 0.9.0
  description: Architecture draft. Representative routes only; per-service exhaustive contracts and legal/medical
    sign-off required. A staff tenant header is a request, not an authority token.
servers:
- url: https://api.example.invalid/api/v2
  description: Placeholder, not a deployed service
...
security:
- bearerAuth: []
paths:
  /me/contexts:
    get:
      summary: List current verified clinic/branch contexts
      tags:
      - Identity
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextList'
        default:
          $ref: '#/components/responses/Error'
      operationId: get_me_contexts
      x-prd-requirements:
      - FR-IAM-02
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error:
      description: Structured error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      additionalProperties: false
      properties:
        error:
          type: object
          additionalProperties: false
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
            details:
              type: object
              additionalProperties: true
          required:
          - code
          - message
          - requestId
      required:
      - error
    ContextList:
      type: object
      additionalProperties: false
      properties:
        contexts:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              clinicId: *id002
              branchIds:
                type: array
                items: *id002
              roles:
                type: array
                items:
                  type: string
              membershipVersion:
                type: integer
            required:
            - clinicId
            - roles
      required:
      - contexts
...